Add ERC721 support to fill-scenarios
