+++ packages/fill-scenarios/src/artifacts.ts
@@ -1,9 +1,11 @@
 import { ContractArtifact } from '@0xproject/sol-compiler';
 
-import * as ERC20Token from './artifacts/ERC20Token.json';
+import * as DummyERC20Token from './artifacts/DummyERC20Token.json';
+import * as DummyERC721Token from './artifacts/DummyERC721Token.json';
 import * as Exchange from './artifacts/Exchange.json';
 
 export const artifacts = {
-    ERC20Token: (ERC20Token as any) as ContractArtifact,
+    DummyERC20Token: (DummyERC20Token as any) as ContractArtifact,
+    DummyERC721Token: (DummyERC721Token as any) as ContractArtifact,
     Exchange: (Exchange as any) as ContractArtifact,
 };
